+++ packages/runtime/src/account.ts
@@ -172,7 +172,8 @@ export class AccountStore implements AccountStoreI {
   }
 
   /**
-   * Creates Asset in account's state
+   * Deploy Asset in account's state
+   * @param assetId Asset Index
    * @param name Asset Name
    * @param asaDef Asset Definitions
    */
@@ -261,7 +262,7 @@ export class AccountStore implements AccountStoreI {
   }
 
   /**
-   * Add application in account's state
+   * Deploy application in account's state
    * check maximum account creation limit
    * @param appID application index
    * @param params SSCDeployment Flags