+++ packages/runtime/test/integration/app-optin.ts
@@ -1,6 +1,5 @@
 import { assert } from "chai";
 
-import { getProgram } from "../../src";
 import { AccountStore, Runtime } from "../../src/index";
 import { ALGORAND_ACCOUNT_MIN_BALANCE } from "../../src/lib/constants";
 import { AppDeploymentFlags } from "../../src/types";
@@ -14,13 +13,13 @@ describe("Algorand Smart Contracts - Stateful Counter example", function () {
   const john = new AccountStore(minBalance + fee);
 
   let runtime: Runtime;
-  let approvalProgram: string;
-  let clearProgram: string;
+  let approvalProgramFileName: string;
+  let clearProgramFileName: string;
   let appID: number;
   let creationFlags: AppDeploymentFlags;
   this.beforeAll(function () {
     runtime = new Runtime([alice, john]); // setup test
-    clearProgram = getProgram('clear.teal');
+    clearProgramFileName = 'clear.teal';
 
     creationFlags = {
       sender: john.account,
@@ -34,9 +33,9 @@ describe("Algorand Smart Contracts - Stateful Counter example", function () {
   const key = "counter";
 
   it("should opt-in to app successfully and update local state", function () {
-    // create new app
-    approvalProgram = getProgram('accept-optin.teal');
-    appID = runtime.addApp(creationFlags, {}, approvalProgram, clearProgram).appID;
+    // deploy new app
+    approvalProgramFileName = 'accept-optin.teal';
+    appID = runtime.deployApp(approvalProgramFileName, clearProgramFileName, creationFlags, {}).appID;
 
     // opt-in (should be accepted)
     assert.doesNotThrow(() => runtime.optInToApp(john.address, appID, {}, {}));
@@ -48,9 +47,9 @@ describe("Algorand Smart Contracts - Stateful Counter example", function () {
   });
 
   it("should reject opt-in to app", function () {
-    // create new app
-    approvalProgram = getProgram('reject-optin.teal');
-    appID = runtime.addApp(creationFlags, {}, approvalProgram, clearProgram).appID;
+    // deploy new app
+    approvalProgramFileName = 'reject-optin.teal';
+    appID = runtime.deployApp(approvalProgramFileName, clearProgramFileName, creationFlags, {}).appID;
 
     // verify local state not present BEFORE optin
     assert.isUndefined(alice.appsLocalState.get(appID));